Extract Rituals Perfume Genie DataUpdateCoordinator into module
