+++ homeassistant/components/rituals_perfume_genie/__init__.py
@@ -1,18 +1,15 @@
 """The Rituals Perfume Genie integration."""
-from datetime import timedelta
-import logging
-
 import aiohttp
-from pyrituals import Account, Diffuser
+from pyrituals import Account
 
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.const import Platform
 from homeassistant.core import HomeAssistant
 from homeassistant.exceptions import ConfigEntryNotReady
 from homeassistant.helpers.aiohttp_client import async_get_clientsession
-from homeassistant.helpers.update_coordinator import DataUpdateCoordinator
 
 from .const import ACCOUNT_HASH, COORDINATORS, DEVICES, DOMAIN
+from .coordinator import RitualsDataUpdateCoordinator
 
 PLATFORMS = [
     Platform.BINARY_SENSOR,
@@ -21,10 +18,6 @@ PLATFORMS = [
     Platform.SENSOR,
     Platform.SWITCH,
 ]
-
-_LOGGER = logging.getLogger(__name__)
-
-UPDATE_INTERVAL = timedelta(minutes=2)
 
 
 async def async_setup_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
@@ -63,21 +56,3 @@ async def async_unload_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
         hass.data[DOMAIN].pop(entry.entry_id)
 
     return unload_ok
-
-
-class RitualsDataUpdateCoordinator(DataUpdateCoordinator[None]):
-    """Class to manage fetching Rituals Perfume Genie device data from single endpoint."""
-
-    def __init__(self, hass: HomeAssistant, device: Diffuser) -> None:
-        """Initialize global Rituals Perfume Genie data updater."""
-        self._device = device
-        super().__init__(
-            hass,
-            _LOGGER,
-            name=f"{DOMAIN}-{device.hublot}",
-            update_interval=UPDATE_INTERVAL,
-        )
-
-    async def _async_update_data(self) -> None:
-        """Fetch data from Rituals."""
-        await self._device.update_data()
