+++ homeassistant/components/rituals_perfume_genie/const.py
@@ -1,7 +1,12 @@
 """Constants for the Rituals Perfume Genie integration."""
+
+from datetime import timedelta
+
 DOMAIN = "rituals_perfume_genie"
 
 ACCOUNT_HASH = "account_hash"
 
 COORDINATORS = "coordinators"
 DEVICES = "devices"
+
+UPDATE_INTERVAL = timedelta(minutes=2)
